GH-1692 Add javadocs integration test
