+++ reposilite-backend/src/integration/kotlin/com/reposilite/javadocs/specification/JavadocsIntegrationSpecification.kt
@@ -1,0 +1,5 @@
+package com.reposilite.javadocs.specification
+
+import com.reposilite.specification.ReposiliteSpecification
+
+internal abstract class JavadocsIntegrationSpecification : ReposiliteSpecification()
